+++ GeoDataServer/src/main/java/cn/edu/whu/gds/service/TileService.java
@@ -1,5 +1,6 @@
 package cn.edu.whu.gds.service;
 
+import cn.edu.whu.gds.bean.vo.Response;
 import io.minio.errors.*;
 import org.imintel.mbtiles4j.MBTilesReadException;
 import org.springframework.stereotype.Service;
